[iOS] Shift AIChat Obj-C wrappers into ai_chat component
