+++ components/ai_chat/ios/browser/DEPS
@@ -1,0 +1,5 @@
+include_rules = [
+  "+components/prefs",
+  "+ios/web/public",
+  "+services/network/public",
+]
