+++ components/ai_chat/ios/browser/conversation_client.h
@@ -3,8 +3,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#ifndef BRAVE_IOS_BROWSER_API_AI_CHAT_CONVERSATION_CLIENT_H_
-#define BRAVE_IOS_BROWSER_API_AI_CHAT_CONVERSATION_CLIENT_H_
+#ifndef BRAVE_COMPONENTS_AI_CHAT_IOS_BROWSER_CONVERSATION_CLIENT_H_
+#define BRAVE_COMPONENTS_AI_CHAT_IOS_BROWSER_CONVERSATION_CLIENT_H_
 
 #include <string>
 #include <vector>
@@ -64,4 +64,4 @@ class ConversationClient : public mojom::ConversationUI,
 
 }  // namespace ai_chat
 
-#endif  // BRAVE_IOS_BROWSER_API_AI_CHAT_CONVERSATION_CLIENT_H_
+#endif  // BRAVE_COMPONENTS_AI_CHAT_IOS_BROWSER_CONVERSATION_CLIENT_H_
