+++ components/ai_chat/ios/browser/headers.gni
@@ -3,13 +3,13 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at https://mozilla.org/MPL/2.0/.
 
-ai_chat_public_headers = [
-  "//brave/ios/browser/api/ai_chat/ai_chat.h",
-  "//brave/ios/browser/api/ai_chat/ai_chat_delegate.h",
+ai_chat_component_public_headers = [
+  "//brave/components/ai_chat/ios/browser/ai_chat.h",
+  "//brave/components/ai_chat/ios/browser/ai_chat_delegate.h",
   "$root_gen_dir/brave/components/ai_chat/core/common/mojom/ios/ai_chat.mojom.objc.h",
   "$root_gen_dir/brave/components/ai_chat/core/common/mojom/ios/tab_tracker.mojom.objc.h",
   "$root_gen_dir/brave/components/ai_chat/core/common/mojom/ios/untrusted_frame.mojom.objc.h",
 ]
 
-ai_chat_public_deps =
-    [ "//brave/ios/browser/api/ai_chat:ai_chat_mojom_wrappers" ]
+ai_chat_component_public_deps =
+    [ "//brave/components/ai_chat/ios/common:mojom" ]
