+++ ios/BUILD.gn
@@ -7,8 +7,8 @@ import("//brave/build/config.gni")
 import("//brave/build/ios/config.gni")
 import("//brave/build/ios/create_xcframework.gni")
 import("//brave/build/mac/tweak_info_plist.gni")
+import("//brave/components/ai_chat/ios/browser/headers.gni")
 import("//brave/ios/browser/api/ads/headers.gni")
-import("//brave/ios/browser/api/ai_chat/headers.gni")
 import("//brave/ios/browser/api/brave_domains/headers.gni")
 import("//brave/ios/browser/api/brave_rewards/headers.gni")
 import("//brave/ios/browser/api/brave_services_key/headers.gni")
@@ -98,7 +98,7 @@ brave_core_public_headers = [
 ]
 
 brave_core_public_headers += ads_public_headers
-brave_core_public_headers += ai_chat_public_headers
+brave_core_public_headers += ai_chat_component_public_headers
 brave_core_public_headers += brave_domains_public_headers
 brave_core_public_headers += brave_shields_public_headers
 brave_core_public_headers += browser_api_url_sanitizer_public_headers
@@ -187,7 +187,7 @@ ios_framework_bundle("brave_core_ios_framework") {
   ]
 
   deps += ads_public_deps
-  deps += ai_chat_public_deps
+  deps += ai_chat_component_public_deps
   deps += brave_wallet_public_deps
   deps += rewards_public_deps
   deps += skus_public_deps
