+++ ios/app/BUILD.gn
@@ -10,7 +10,7 @@ import("//ios/public/provider/chrome/browser/build_config.gni")
 
 source_set("app") {
   configs += [
-    "//brave/ios/browser/api/ai_chat:mojom_header_config",
+    "//brave/components/ai_chat/ios/browser:mojom_header_config",
     "//brave/ios/browser/api/webcompat_reporter:mojom_header_config",
     "//brave/ios/browser/api/brave_wallet:mojom_header_config",
     "//brave/ios/browser/api/skus:mojom_header_config",
@@ -31,6 +31,8 @@ source_set("app") {
   deps = [
     "//base",
     "//brave/components/ai_chat/core/browser",
+    "//brave/components/ai_chat/ios/browser",
+    "//brave/components/ai_chat/ios/common:mojom",
     "//brave/components/brave_component_updater/browser",
     "//brave/components/brave_sync:constants",
     "//brave/components/brave_user_agent/browser",
@@ -44,9 +46,8 @@ source_set("app") {
     "//brave/components/variations",
     "//brave/ios/app/resources",
     "//brave/ios/browser",
+    "//brave/ios/browser/ai_chat",
     "//brave/ios/browser/api/ads",
-    "//brave/ios/browser/api/ai_chat",
-    "//brave/ios/browser/api/ai_chat:ai_chat_mojom_wrappers",
     "//brave/ios/browser/api/bookmarks",
     "//brave/ios/browser/api/brave_rewards",
     "//brave/ios/browser/api/brave_shields",
