+++ ios/app/brave_profile_controller.mm
@@ -7,9 +7,11 @@
 
 #include "base/check.h"
 #include "base/memory/raw_ptr.h"
+#include "brave/components/ai_chat/ios/browser/ai_chat+private.h"
+#include "brave/components/ai_chat/ios/browser/ai_chat_delegate.h"
 #include "brave/components/ntp_background_images/browser/ntp_background_images_service.h"
-#include "brave/ios/browser/api/ai_chat/ai_chat+private.h"
-#include "brave/ios/browser/api/ai_chat/ai_chat_delegate.h"
+#include "brave/ios/browser/ai_chat/ai_chat_service_factory.h"
+#include "brave/ios/browser/ai_chat/model_service_factory.h"
 #include "brave/ios/browser/api/bookmarks/brave_bookmarks_api+private.h"
 #include "brave/ios/browser/api/brave_stats/brave_stats+private.h"
 #include "brave/ios/browser/api/brave_wallet/brave_wallet_api+private.h"
@@ -66,6 +68,7 @@
 #include "ios/web_view/internal/cwv_web_view_configuration_internal.h"
 #include "ios/web_view/internal/web_view_browser_state.h"
 #include "ios/web_view/internal/web_view_download_manager.h"
+#include "services/network/public/cpp/shared_url_loader_factory.h"
 
 #if BUILDFLAG(IOS_CREDENTIAL_PROVIDER_ENABLED)
 #include "ios/chrome/browser/credential_provider/model/credential_provider_service_factory.h"
@@ -311,7 +314,14 @@
 }
 
 - (AIChat*)aiChatAPIWithDelegate:(id<AIChatDelegate>)delegate {
-  return [[AIChat alloc] initWithProfileIOS:_profile delegate:delegate];
+  auto* modelService = ai_chat::ModelServiceFactory::GetForProfile(_profile);
+  auto* service = ai_chat::AIChatServiceFactory::GetForProfile(_profile);
+  return [[AIChat alloc]
+      initWithAIChatService:service
+               modelService:modelService
+               profilePrefs:_profile->GetPrefs()
+      sharedURLoaderFactory:_profile->GetSharedURLLoaderFactory()
+                   delegate:delegate];
 }
 
 - (DefaultHostContentSettings*)defaultHostContentSettings {
