+++ ios/browser/ai_chat/ai_chat_service_factory.h
@@ -3,12 +3,13 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#ifndef BRAVE_IOS_BROWSER_API_AI_CHAT_MODEL_SERVICE_FACTORY_H_
-#define BRAVE_IOS_BROWSER_API_AI_CHAT_MODEL_SERVICE_FACTORY_H_
+#ifndef BRAVE_IOS_BROWSER_AI_CHAT_AI_CHAT_SERVICE_FACTORY_H_
+#define BRAVE_IOS_BROWSER_AI_CHAT_AI_CHAT_SERVICE_FACTORY_H_
 
 #include <memory>
 
 #include "base/no_destructor.h"
+#include "brave/components/ai_chat/core/browser/ai_chat_metrics.h"
 #include "ios/chrome/browser/shared/model/profile/profile_keyed_service_factory_ios.h"
 
 class ProfileIOS;
@@ -21,22 +22,21 @@ class NoDestructor;
 
 namespace ai_chat {
 
-class ModelService;
+class AIChatService;
 
-class ModelServiceFactory : public ProfileKeyedServiceFactoryIOS {
+class AIChatServiceFactory : public ProfileKeyedServiceFactoryIOS {
  public:
-  static ModelService* GetForProfile(ProfileIOS* profile);
-  static ModelService* GetForProfileIfExists(ProfileIOS* profile);
-  static ModelServiceFactory* GetInstance();
+  static AIChatService* GetForProfile(ProfileIOS* profile);
+  static AIChatServiceFactory* GetInstance();
 
-  ModelServiceFactory(const ModelServiceFactory&) = delete;
-  ModelServiceFactory& operator=(const ModelServiceFactory&) = delete;
+  AIChatServiceFactory(const AIChatServiceFactory&) = delete;
+  AIChatServiceFactory& operator=(const AIChatServiceFactory&) = delete;
 
  private:
-  friend class base::NoDestructor<ModelServiceFactory>;
+  friend class base::NoDestructor<AIChatServiceFactory>;
 
-  ModelServiceFactory();
-  ~ModelServiceFactory() override;
+  AIChatServiceFactory();
+  ~AIChatServiceFactory() override;
 
   // ProfileKeyedServiceFactoryIOS implementation.
   std::unique_ptr<KeyedService> BuildServiceInstanceFor(
@@ -44,4 +44,4 @@ class ModelServiceFactory : public ProfileKeyedServiceFactoryIOS {
 };
 }  // namespace ai_chat
 
-#endif  // BRAVE_IOS_BROWSER_API_AI_CHAT_MODEL_SERVICE_FACTORY_H_
+#endif  // BRAVE_IOS_BROWSER_AI_CHAT_AI_CHAT_SERVICE_FACTORY_H_
