+++ ios/browser/ai_chat/model_service_factory.mm
@@ -3,7 +3,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "brave/ios/browser/api/ai_chat/model_service_factory.h"
+#include "brave/ios/browser/ai_chat/model_service_factory.h"
 
 #include "base/no_destructor.h"
 #include "brave/components/ai_chat/core/browser/model_service.h"
