+++ ios/browser/profile/model/BUILD.gn
@@ -14,7 +14,7 @@ source_set("model") {
   deps = [
     "//base",
     "//brave/components/brave_account:features",
-    "//brave/ios/browser/api/ai_chat",
+    "//brave/ios/browser/ai_chat",
     "//brave/ios/browser/api/webcompat_reporter",
     "//brave/ios/browser/brave_account",
     "//brave/ios/browser/brave_ads",
