+++ ios/browser/profile/model/brave_keyed_service_factories.mm
@@ -6,8 +6,8 @@
 #include "brave/ios/browser/profile/model/brave_keyed_service_factories.h"
 
 #include "brave/components/brave_account/features.h"
-#include "brave/ios/browser/api/ai_chat/ai_chat_service_factory.h"
-#include "brave/ios/browser/api/ai_chat/model_service_factory.h"
+#include "brave/ios/browser/ai_chat/ai_chat_service_factory.h"
+#include "brave/ios/browser/ai_chat/model_service_factory.h"
 #include "brave/ios/browser/api/webcompat_reporter/webcompat_reporter_service_factory.h"
 #include "brave/ios/browser/brave_account/brave_account_service_factory_ios.h"
 #include "brave/ios/browser/brave_ads/ads_service_factory_ios.h"
